simplified app.tsx by moving loadCoursebag() into the useCoursebag() hook, and simplifying html
