+++ site/src/hooks/coursebag.ts
@@ -1,7 +1,8 @@
 import { useCallback, useEffect } from 'react';
 import { useAppDispatch, useAppSelector } from '../store/hooks';
-import { addCourseToBagState, removeCourseFromBagState } from '../store/slices/coursebagSlice';
+import { setCoursebag, addCourseToBagState, removeCourseFromBagState } from '../store/slices/coursebagSlice';
 import trpc from '../trpc';
+import { searchAPIResults } from '../helpers/util';
 import { useIsLoggedIn } from './isLoggedIn';
 import { CourseGQLData } from '../types/types';
 
@@ -44,5 +45,13 @@ export function useCoursebag() {
     [addCourseToBag, coursebag, removeCourseFromBag],
   );
 
-  return { coursebag: coursebag ?? [], addCourseToBag, removeCourseFromBag, toggleBookmark };
+  const loadCoursebag = useCallback(async () => {
+    const courseIds = isLoggedIn
+      ? await trpc.savedCourses.get.query()
+      : JSON.parse(localStorage.getItem('coursebag') ?? '[]');
+    const coursebag = await searchAPIResults('courses', courseIds);
+    dispatch(setCoursebag(Object.values(coursebag)));
+  }, [dispatch, isLoggedIn]);
+
+  return { coursebag: coursebag ?? [], addCourseToBag, removeCourseFromBag, toggleBookmark, loadCoursebag };
 }
